Fixed offset augment for patches dataset

+++ dh_segment_torch/data/datasets/patches_dataset.py
@@ -287,8 +287,8 @@ def sample_to_patche_samples(
         offsets = np.random.rand(2)
         h, w = sample["image"].shape[:2]
 
-        offset_h = (offsets[0] * h) // 2
-        offset_w = (offsets[1] * w) // 2
+        offset_h = int((offsets[0] * h) // 2)
+        offset_w = int((offsets[1] * w) // 2)
     else:
         offset_h = 0
         offset_w = 0
